requirements.txt
absl-py==0.13.0
alabaster==0.7.12
alembic==1.4.1
antlr4-python3-runtime==4.8
anyio==3.2.0
argcomplete==1.12.3
Babel==2.9.1
boto3==1.17.97
botocore==1.20.97
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
cloudpickle==1.6.0
cycler==0.10.0
databricks-cli==0.14.3
decorator==4.4.2
Deprecated==1.2.12
dill==0.3.4
dnspython==1.16.0
docker==5.0.0
docutils==0.17.1
entrypoints==0.3
eventlet==0.31.0
fiftyone==0.9.4
fiftyone-brain==0.6.1
fiftyone-db==0.2.1
Flask==2.0.1
flatbuffers==2.0
future==0.18.2
gitdb==4.0.7
GitPython==3.1.18
glob2==0.7
google-auth==1.31.0
google-auth-oauthlib==0.4.4
greenlet==1.1.0
grpcio==1.38.0
gunicorn==20.1.0
h11==0.12.0
httpcore==0.13.6
httpx==0.18.2
hydra-core==1.1.0
idna==2.10
imageio==2.9.0
imagesize==1.2.0
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
kaleido==0.2.1
kiwisolver==1.3.1
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
mlflow==1.18.0
mongoengine==0.20.0
motor==2.4.0
ndjson==0.3.1
networkx==2.5.1
numpy==1.20.3
oauthlib==3.1.1
omegaconf==2.1.0
onnx==1.9.0
onnxruntime==1.8.0
opencv-python-headless==4.4.0.46
packaging==20.9
pandas==1.2.4
patool==1.12
Pillow==8.2.0
plotly==4.14.3
pprintpp==0.4.0
prometheus-client==0.11.0
prometheus-flask-exporter==0.18.2
protobuf==3.17.3
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.9.0
pymongo==3.11.4
pyparsing==2.4.7
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
querystring-parser==1.2.4
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
rfc3986==1.5.0
rsa==4.7.2
s3transfer==0.4.2
scikit-image==0.18.1
scikit-learn==0.24.2
scipy==1.6.3
six==1.16.0
smmap==4.0.0
sniffio==1.2.0
snowballstemmer==2.1.0
sortedcontainers==2.4.0
Sphinx==4.0.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.18
sqlparse==0.4.1
tabulate==0.8.9
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
threadpoolctl==2.1.0
tifffile==2021.6.14
torch==1.9.0
torchvision==0.10.0
tornado==6.1
typing-extensions==3.10.0.0
tzlocal==2.1
universal-analytics-python3==1.1.1
urllib3==1.26.5
voxel51-eta==0.4.1
websocket-client==1.1.0
Werkzeug==2.0.1
wrapt==1.12.1
xmltodict==0.12.0
